// src/graphql-types.d.ts
export interface PageInfo {
    hasNextPage: boolean;
}
// ...
export interface Node {
    id: string;
    parent: Node | null;
    children: Array<Node> | null;
}
// ...
export interface ImageSharp extends Node {
    id: string;
    parent: Node | null;
    children: Array<Node> | null;
    internal: internal_14 | null;
    original: ImageSharpOriginal | null;
    resolutions: ImageSharpResolutions | null;
    sizes: ImageSharpSizes | null;
    fluid: ImageSharpfluid | null;
    responsiveSizes: ImageSharpResponsiveSizes | null;
    resize: ImageSharpResize | null;
}
// ...
export interface internal_14 {
    contentDigest: string | null;
    type: string | null;
    owner: string | null;
}

export interface ImageSharpOriginal {
    width: number | null;
    height: number | null;
    src: string | null;
}
// ...
export interface ImageSharpResolutions {
    base64: string | null;
    tracedSVG: string | null;
    aspectRatio: number | null;
    width: number | null;
    height: number | null;
    src: string | null;
    srcSet: string | null;
    srcWebp: string | null;
    srcSetWebp: string | null;
    originalName: string | null;
}

export interface ImageSharpSizes {
    base64: string | null;
    tracedSVG: string | null;
    aspectRatio: number | null;
    src: string | null;
    srcSet: string | null;
    srcWebp: string | null;
    srcSetWebp: string | null;
    sizes: string | null;
    originalImg: string | null;
    originalName: string | null;
}

export interface ImageSharpfluid {
    base64: string | null;
    aspectRatio: number | null;
    width: number | null;
    height: number | null;
    src: string | null;
    srcSet: string | null;
    originalName: string | null;
}

export interface ImageSharpResponsiveSizes {
    base64: string | null;
    aspectRatio: number | null;
    src: string | null;
    srcSet: string | null;
    sizes: string | null;
    originalImg: string | null;
    originalName: string | null;
}

export interface ImageSharpResize {
    src: string | null;
    tracedSVG: string | null;
    width: number | null;
    height: number | null;
    aspectRatio: number | null;
    originalName: string | null;
}

export interface MarkdownRemark extends Node {
    id: string;
    parent: Node | null;
    children: Array<Node> | null;
    internal: internal_15 | null;
    frontmatter: frontmatter_2 | null;
    fileAbsolutePath: string | null;
    fields: fields_2 | null;
    html: string | null;
    excerpt: string | null;
    headings: Array<MarkdownHeading> | null;
    timeToRead: number | null;
    tableOfContents: string | null;
    wordCount: wordCount | null;
}
// ...
export interface MarkdownRemarkConnection {
    pageInfo: PageInfo;
    edges: Array<MarkdownRemarkEdge> | null;
    totalCount: number | null;
    distinct: Array<string> | null;
    group: Array<markdownRemarkGroupConnectionConnection> | null;
}

export interface DistinctMarkdownRemarkConnectionArgs {
    field: markdownRemarkDistinctEnum | null;
}

export interface GroupMarkdownRemarkConnectionArgs {
    skip: number | null;
    limit: number | null;
    field: markdownRemarkGroupEnum | null;
}

export interface MarkdownRemarkEdge {
    node: MarkdownRemark | null;
    next: MarkdownRemark | null;
    previous: MarkdownRemark | null;
}
// ...
export interface markdownRemarkGroupConnectionConnection {
    pageInfo: PageInfo;
    edges: Array<markdownRemarkGroupConnectionEdge> | null;
    field: string | null;
    fieldValue: string | null;
    totalCount: number | null;
}

export interface markdownRemarkGroupConnectionEdge {
    node: MarkdownRemark | null;
    next: MarkdownRemark | null;
    previous: MarkdownRemark | null;
}